vim-yankstack
Yankstack.vim is a lightweight implementation of the Emacs 'kill ring' for
Vim. It allows you to yank and delete things without worrying about losing the
text that you yanked previously. It effectively turns your default register
into a stack, and lets you cycle through the items in the stack after doing a
paste.